requirements.txt
matplotlib~=3.4.3
numpy~=1.20.3
Pillow~=8.4.0
torch
torchvision~=0.12.0
tqdm~=4.62.3
wandb~=0.12.11
einops~=0.4.1
imageio~=2.9.0
scikit-image~=0.18.3
colour~=0.1.5